<<Negocios>> Alta de eventos algun cambio mas por el float y double

+++ GeoRed/source/Negocios/source/negocios/GestionUsuarios.java
@@ -115,7 +115,7 @@ public interface GestionUsuarios {
 	 *            Identificador del usuario.
 	 * @return Lista de notficaciones a mostrar.
 	 */
-	public List<Notificacion> getNotificaciones(final int idUsuario, final float latitud, final float longitud, final float distancia) throws EntidadNoExiste;
+	public List<Notificacion> getNotificaciones(final int idUsuario, final double latitud, final double longitud, final double distancia) throws EntidadNoExiste;
 
 	/**
 	 * Permite al usuario realizar una compra de una ofera
